角色菜单资源查询

+++ src/main/java/top/baozoulolw/exam/service/impl/ResourceServiceImpl.java
@@ -6,7 +6,9 @@ import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.stereotype.Service;
 import top.baozoulolw.exam.common.Result;
 import top.baozoulolw.exam.dao.ResourceDao;
+import top.baozoulolw.exam.dao.RoleResourceDao;
 import top.baozoulolw.exam.entity.Resource;
+import top.baozoulolw.exam.entity.RoleResource;
 import top.baozoulolw.exam.service.ResourceService;
 
 import java.util.List;
@@ -16,6 +18,9 @@ public class ResourceServiceImpl extends ServiceImpl<ResourceDao, Resource> impl
 
     @javax.annotation.Resource
     private ResourceDao resourceDao;
+
+    @javax.annotation.Resource
+    private RoleResourceDao roleResourceDao;
 
     @Override
     public Result<List<Resource>> getResById(Long id) {
@@ -54,6 +59,12 @@ public class ResourceServiceImpl extends ServiceImpl<ResourceDao, Resource> impl
         Page<Resource> resourcePage = resourceDao.selectPage(new Page<>(1, 1), wrapper);
         if(resourcePage.getRecords().size() > 0){
             return Result.fail("该资源已有子资源关联，无法删除");
+        }
+        QueryWrapper<RoleResource> wrapper1 = new QueryWrapper<>();
+        wrapper1.eq("resource_id", id);
+        Page<RoleResource> roleResourcePage = roleResourceDao.selectPage(new Page<>(1, 1), wrapper1);
+        if(roleResourcePage.getRecords().size() > 0){
+            return Result.fail("该资源已有角色与其关联，无法删除");
         }
         resourceDao.deleteById(id);
         return Result.success();
